favorites через cookie

+++ app/Http/Controllers/Auth/ClientController.php
@@ -85,22 +85,64 @@ class ClientController extends Controller
 
     public function addFavorite(Request $request)
     {
-        if (!is_null($Client = Auth::guard('client')->user())) {
+        if (ClientFavorite::$Type == 0) {
+            if (!is_null($Client = Auth::guard('client')->user())) {
 
-            $ClientFavorite = ClientFavorite::where("shop_item_id", $request->shop_item_id)
-                                                ->where("client_id", $Client->id)->first();
-            if (is_null($ClientFavorite)) {
-                $ClientFavorite = new ClientFavorite();
-                $ClientFavorite->shop_item_id = $request->shop_item_id;
-                $ClientFavorite->client_id = $Client->id;
-                $ClientFavorite->save();
+                $ClientFavorite = ClientFavorite::where("shop_item_id", $request->shop_item_id)
+                                                    ->where("client_id", $Client->id)->first();
+                if (is_null($ClientFavorite)) {
+                    $ClientFavorite = new ClientFavorite();
+                    $ClientFavorite->shop_item_id = $request->shop_item_id;
+                    $ClientFavorite->client_id = $Client->id;
+                    $ClientFavorite->save();
+    
+                } else {
+                    $ClientFavorite->delete();
+                    
+                }
+    
+                return response()->json(["count" => ClientFavorite::where("client_id", $Client->id)->count()]);
+            }
+        } else if (ClientFavorite::$Type == 1) {
 
+            $count = isset($_COOKIE["favorites"]) ? count($_COOKIE["favorites"]) : 0;
+
+            if (isset($_COOKIE["favorites"][$request->shop_item_id])) {
+                unset($_COOKIE["favorites"][$request->shop_item_id]);
+                setcookie('favorites['. $request->shop_item_id .']', null, -1, '/');
+                $count = $count - 1;
             } else {
-                $ClientFavorite->delete();
-                
+                setcookie("favorites[". $request->shop_item_id ."]", 1, time() + ClientFavorite::$CookieTime, "/");
+                $count = $count + 1;
             }
 
-            return response()->json(["count" => ClientFavorite::where("client_id", $Client->id)->count()]);
+            return response()->json(["count" => $count]);            
+        } 
+
+    }
+
+    public static function getCookieFavorites()
+    {
+        $Result = [];
+        if (isset($_COOKIE["favorites"])) {
+            foreach ($_COOKIE["favorites"] as $key => $value) {
+                $Result[] = $key;
+            }
         }
+
+        return $Result;
+    }
+
+    public static function getCookieFavoritesCount()
+    {
+        return count(self::getCookieFavorites());
+    }
+
+    public function cookieFavorites()
+    {
+        $items = self::getCookieFavorites();
+        return view('client.cookie-favorite', [
+            'shopItems' => count($items) > 0 ? ShopItem::whereIn("id", $items)->where("active", 1)->paginate(self::$items_on_page) : false,
+        ]);
     }
 }
